Organized code into modules
