+++ frontend/src/components/sitzverteilung/SitzverteilungChart.tsx
