+++ frontend/src/components/sitzverteilung/SitzverteilungOverview.tsx
