+++ frontend/src/components/sitzverteilung/SitzverteilungTable.tsx
