+++ frontend/src/components/stimmkreise/StimmkreisChart.tsx
