+++ frontend/src/components/stimmkreise/StimmkreisTable.tsx
